refactor: selection background extract method
